+++ apps/blog-app/vite.config.ts
@@ -19,6 +19,13 @@ export default defineConfig(() => {
     plugins: [
       analog({
         static: true,
+        vite: {
+          experimental: {
+            supportAnalogFormat: true,
+          },
+        },
+        additionalPagesDirs: ['/libs/shared/feature'],
+        additionalContentDirs: ['/libs/shared/feature/src/content'],
         content: {
           highlighter: 'shiki',
           shikiOptions: {