+++ packages/router/src/lib/route-config.ts
@@ -1,11 +1,12 @@
 import { inject } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import type { Route } from '@angular/router';
+import { injectBaseURL } from '@analogjs/router/tokens';
 import { firstValueFrom } from 'rxjs';
 
 import { RedirectRouteMeta, RouteConfig, RouteMeta } from './models';
 import { ROUTE_META_TAGS_KEY } from './meta-tags';
-import { PAGE_ENDPOINTS, ANALOG_META_KEY } from './endpoints';
+import { ANALOG_PAGE_ENDPOINTS, ANALOG_META_KEY } from './endpoints';
 
 export function toRouteConfig(routeMeta: RouteMeta | undefined): RouteConfig {
   if (routeMeta && isRedirectRouteMeta(routeMeta)) {
@@ -30,15 +31,24 @@ export function toRouteConfig(routeMeta: RouteMeta | undefined): RouteConfig {
   routeConfig.resolve = {
     ...routeConfig.resolve,
     load: async (route) => {
+      const baseUrl = injectBaseURL();
       const routeConfig = route.routeConfig as Route & {
         [ANALOG_META_KEY]: { endpoint: string; endpointKey: string };
       };
 
-      if (PAGE_ENDPOINTS[routeConfig[ANALOG_META_KEY].endpointKey]) {
+      if (ANALOG_PAGE_ENDPOINTS[routeConfig[ANALOG_META_KEY].endpointKey]) {
         const { queryParams, fragment: hash, params, parent } = route;
         const segment =
           parent?.url.map((segment) => segment.path).join('/') || '';
-        const url = new URL('', import.meta.env['VITE_ANALOG_PUBLIC_BASE_URL']);
+        const url = new URL(
+          '',
+          import.meta.env['VITE_ANALOG_PUBLIC_BASE_URL'] ||
+            baseUrl ||
+            (typeof window !== 'undefined' && window.location.origin
+              ? window.location.origin
+              : '')
+        );
+
         url.pathname = `${
           url.pathname.endsWith('/') ? url.pathname : url.pathname + '/'
         }api/_analog${routeConfig[ANALOG_META_KEY].endpoint}`;
@@ -50,7 +60,10 @@ export function toRouteConfig(routeMeta: RouteMeta | undefined): RouteConfig {
         });
         url.pathname = url.pathname.replace('**', segment);
 
-        if ((globalThis as any).$fetch) {
+        if (
+          !!import.meta.env['VITE_ANALOG_PUBLIC_BASE_URL'] &&
+          (globalThis as any).$fetch
+        ) {
           return (globalThis as any).$fetch(url.pathname);
         }