+++ packages/vite-plugin-nitro/src/lib/options.ts
@@ -11,6 +11,14 @@ export interface Options {
   entryServer?: string;
   index?: string;
   workspaceRoot?: string;
+  /**
+   * Additional page paths to include
+   */
+  additionalPagesDirs?: string[];
+  /**
+   * Additional API paths to include
+   */
+  additionalAPIDirs?: string[];
 }
 
 export interface PrerenderOptions {